requirements.txt
av
ffmpeg-python
Flask
Flask-Cors
Flask-HTTPAuth
imageio
numpy
Pillow
pillow-heif
PyExifTool
pyiqa
python-dotenv
torch==2.5.0.dev20240827
torchvision==0.20.0.dev20240827